refactor : 스케줄링으로 부터 트랜잭션 분리 & @UtilityClass적용
