+++ src/main/java/com/flab/offcoupon/controller/HomeController.java
@@ -2,12 +2,10 @@ package com.flab.offcoupon.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
 
 @Slf4j
-@RequestMapping("/")
 @RestController
 public class HomeController {
