+++ src/main/java/com/flab/offcoupon/util/DateTimeUtils.java
@@ -1,13 +1,12 @@
 package com.flab.offcoupon.util;
 
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
+import lombok.experimental.UtilityClass;
 
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 
-@NoArgsConstructor (access = AccessLevel.PRIVATE)
+@UtilityClass
 public class DateTimeUtils {
 
     public static LocalDateTime nowFromZone() {
